git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@68579 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/common/VBoxGuestLib/VBoxGuestR3Lib.cpp
@@ -428,7 +428,7 @@ int vbglR3DoIOCtlRaw(uintptr_t uFunction, PVBGLREQHDR pHdr, size_t cbReq)
 #else
     if (g_File != NIL_RTFILE)
     {
-        if (RT_LIKELY(ioctl((int)g_File, uFunction, pHdr) >= 0))
+        if (RT_LIKELY(ioctl((int)(intptr_t)g_File, uFunction, pHdr) >= 0))
             return VINF_SUCCESS;
         return RTErrConvertFromErrno(errno);
     }
